+++ src/main/java/youth/week6/member/jwt/service/CustomAuthenticationProvider.java
@@ -1,0 +1,51 @@
+package youth.week6.member.jwt.service;
+
+import org.springframework.security.authentication.AuthenticationProvider;
+import org.springframework.security.authentication.BadCredentialsException;
+import org.springframework.security.core.Authentication;
+import org.springframework.security.core.AuthenticationException;
+import org.springframework.security.crypto.password.PasswordEncoder;
+import org.springframework.stereotype.Component;
+import youth.week6.member.jwt.dto.CustomAuthenticationToken;
+import youth.week6.member.jwt.dto.CustomUserDetails;
+
+@Component
+public class CustomAuthenticationProvider implements AuthenticationProvider {
+
+    private final CustomUserDetailsService customUserDetailsService;
+    private final PasswordEncoder passwordEncoder;
+
+    public CustomAuthenticationProvider(CustomUserDetailsService customUserDetailsService,
+        PasswordEncoder passwordEncoder) {
+        this.customUserDetailsService = customUserDetailsService;
+        this.passwordEncoder = passwordEncoder;
+    }
+
+    @Override
+    public CustomAuthenticationToken authenticate(Authentication authentication)
+        throws AuthenticationException {
+        String username = authentication.getName();
+        String password = (String) authentication.getCredentials();
+
+        CustomUserDetails user = customUserDetailsService.loadUserByUsername(username);
+        if (user == null) {
+            throw new BadCredentialsException("username is not found. username=" + username);
+        }
+
+        if (!this.passwordEncoder.matches(password, user.getPassword())) {
+            throw new BadCredentialsException("password is not matched");
+        }
+
+        return new CustomAuthenticationToken(
+            user.getUsername(),
+            user.getPassword(),
+            user.getMemberId(),
+            user.getAuthorities()
+        );
+    }
+
+    @Override
+    public boolean supports(Class<?> authentication) {
+        return CustomAuthenticationToken.class.isAssignableFrom(authentication);
+    }
+}