+++ src/main/java/youth/week6/member/member/entity/MemberRoles.java
@@ -1,0 +1,7 @@
+package youth.week6.member.member.entity;
+
+public enum MemberRoles {
+    USER,
+    PARTICIPANT,
+    ORGANIZER;
+}