+++ src/main/java/youth/week6/member/member/entity/embeded/AuthenticationInfo.java
@@ -1,11 +1,15 @@
 package youth.week6.member.member.entity.embeded;
 
+import java.util.EnumSet;
 import java.util.Objects;
 import javax.persistence.Column;
+import javax.persistence.Convert;
 import javax.persistence.Embeddable;
 import lombok.AccessLevel;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
+import youth.week6.member.member.entity.MemberRoles;
+import youth.week6.member.member.entity.utils.MemberRolesConverter;
 
 @Getter
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
@@ -17,9 +21,22 @@ public class AuthenticationInfo {
     @Column(nullable = false)
     private String password;
 
+    @Convert(converter = MemberRolesConverter.class)
+    private EnumSet<MemberRoles> roles = EnumSet.noneOf(MemberRoles.class);
+
     public AuthenticationInfo(String identification, String password) {
         this.identification = identification;
         this.password = password;
+        this.roles = EnumSet.noneOf(MemberRoles.class);
+        roles.add(MemberRoles.USER);
+    }
+
+    public void addRole(MemberRoles memberRoles) {
+        roles.add(memberRoles);
+    }
+
+    public void removeRole(MemberRoles memberRoles) {
+        roles.remove(memberRoles);
     }
 
     @Override